+++ packages/full/index.js
@@ -2,7 +2,3 @@ export * from "@croquet/worldcore-kernel";
 export * from "@croquet/worldcore-three";
 export * from "@croquet/worldcore-widget";
 export * from "@croquet/worldcore-widget2";
-
-// export * from "@croquet/worldcore-rapier";
-// export * from "@croquet/worldcore-widget3";
-// export * from "@croquet/worldcore-widget";