+++ examples/physics/index.js
@@ -11,10 +11,8 @@ App.makeWidgetDock({debug: true, stats: true});
 StartWorldcore({
     appId: 'io.croquet.physics',
     apiKey: '1Mnk3Gf93ls03eu0Barbdzzd3xl1Ibxs7khs8Hon9',
-    name: 'Physics',
-    password: 'password',
     model: MyModelRoot,
     name: App.autoSession(),
+    password: 'password',
     view: MyViewRoot,
-    tps:60
 });