+++ packages/rapier/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+Kinematic rigid bodies
+
 ### Changed
 Upgraded to Rapier 0.11.0
 RapierManager doesn't hold world