+++ packages/rapier/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+Snapshot RigidBody's userData
+
 ### Added
 Kinematic rigid bodies